refactor: Product tree
