+++ src/lib/singleview/vulnerabilities/vulnerability/GeneralSection.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import KeyValues from "$lib/singleview/KeyValue.svelte";
+  import KeyValue from "$lib/singleview/KeyValue.svelte";
 
   export let vulnerability: any;
   const keys: Array<String> = [];
@@ -24,4 +24,4 @@
   }
 </script>
 
-<KeyValues {keys} {values} />
+<KeyValue {keys} {values} />
