+++ src/lib/singleview/vulnerabilities/vulnerability/ID.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import Collapsible from "$lib/Collapsible.svelte";
-  import KeyValues from "$lib/singleview/KeyValue.svelte";
+  import KeyValue from "$lib/singleview/KeyValue.svelte";
   export let vulnerability: any;
 </script>
 
 <Collapsible header="IDs" level="4">
   {#each vulnerability.ids as vulnerabilityID}
-    <KeyValues
+    <KeyValue
       keys={["Systemname", "Text"]}
       values={[vulnerabilityID.system_name, vulnerabilityID.text]}
     />
