+++ src/lib/singleview/vulnerabilities/vulnerability/acknowledgements/Acknowledgement.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import KeyValues from "$lib/singleview/KeyValue.svelte";
+  import KeyValue from "$lib/singleview/KeyValue.svelte";
   import ValueList from "$lib/singleview/ValueList.svelte";
   export let ack: any;
   const keys: Array<String> = [];
@@ -18,7 +18,7 @@
   }
 </script>
 
-<KeyValues {keys} {values} />
+<KeyValue {keys} {values} />
 {#if ack.urls}
   <ValueList label="URLs" values={ack.urls} />
 {/if}
