+++ src/lib/singleview/vulnerabilities/vulnerability/scores/CVSSV2.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Collapsible from "$lib/Collapsible.svelte";
-  import KeyValues from "$lib/singleview/KeyValue.svelte";
+  import KeyValue from "$lib/singleview/KeyValue.svelte";
   export let score: any;
   const keys: Array<String> = [];
   const values: Array<String> = [];
@@ -83,5 +83,5 @@
 </script>
 
 <Collapsible header="CVSS V2" level="5">
-  <KeyValues {keys} {values} />
+  <KeyValue {keys} {values} />
 </Collapsible>
