+++ src/lib/singleview/vulnerabilities/vulnerability/scores/CVSSV3.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Collapsible from "$lib/Collapsible.svelte";
-  import KeyValues from "$lib/singleview/KeyValue.svelte";
+  import KeyValue from "$lib/singleview/KeyValue.svelte";
   export let score: any;
   const keys: Array<String> = [];
   const values: Array<String> = [];
@@ -127,5 +127,5 @@
 </script>
 
 <Collapsible header="CVSS V{score.cvss_v3.version}" level="5">
-  <KeyValues {keys} {values} />
+  <KeyValue {keys} {values} />
 </Collapsible>
